\name{data-sets}

\alias{macro}

\encoding{latin1}

\title{Macroeconomic Time Series Data Sets, 1967M1-2025M7}

\description{

macro contains monthly observations from 1967M1 to 2025M7 for the unemployment rate (the dependent variable), IC (Initial Claims of unemployment insurance), inflation rate (seasonal growth rate of CPI), industrial growth (seasonal growth rate of the industrial production index).r

}


\usage{
data(macro)
}



\value{

    \code{macro} is an object of class \code{timeSeries}.

}


\keyword{datasets}

