\name{fmvar_forecast}

\alias{fmvar_forecast}

\encoding{latin1}

\title{Forecast a FM-VAR System}

\description{Forecast a VAR generated by fmvar_forecast.}

\usage{
fmvar_forecast(output, n.ahead=6)
}

\arguments{

  \item{output}{The output object of fmvar_forecast.}
  \item{n.ahead}{The steps of out-of-sample forecasting.}

}
\details{
This function recursively computes the n.ahead steps of out-of-sample forecasting.
}

\value{

Forecasted values of all endogenous variables.

}

\examples{
data(macro)
out=fmvar(macro,p=1,q=6,v=10,type="trend",ker_fun="parzen")
fmvar_forecast(out,n.ahead=6)

}

\author{
Ho Tsung-wu <tsungwu@ntnu.edu.tw>, College of Management, National Taiwan Normal University.
}

\keyword{cointegration}
